user registration done

+++ base/views.py
@@ -1,3 +1,4 @@
+from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 
 from base.forms import RoomForm
@@ -6,13 +7,19 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 
 
 # Create your views here.
 def loginPage(request):
 
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home_index')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -27,7 +34,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or Password doesnot match!')
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
 
 
@@ -35,6 +42,24 @@ def logoutUser(request):
     logout(request)
     return redirect('home_index')
 
+
+def registerUser(request):
+    form = UserCreationForm()
+    context = {'form':form}
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home_index')
+        else:
+            messages.error(request, 'An error occured!')
+
+        
+    return render(request, 'base/login_register.html', context)
 
 
 def RoomView(request, id):
@@ -58,6 +83,7 @@ def Home(request):
     return render(request, './base/home.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -70,9 +96,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, id):
     room = get_object_or_404(Room, pk=id)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to make changes!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -84,8 +114,13 @@ def updateRoom(request, id):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, id):
     room = get_object_or_404(Room, pk=id)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to make changes!!')
+
     context = {'room':room}
     if request.method == 'POST':
         room.delete()
